refactor: replace legacy risk manager with service
