+++ src/tradingbot/strategies/cross_exchange_arbitrage.py
@@ -8,7 +8,6 @@ from typing import Dict, Optional
 
 from ..adapters.base import ExchangeAdapter
 from ..execution.balance import rebalance_between_exchanges
-from ..risk.manager import RiskManager
 from ..risk.portfolio_guard import GuardConfig, PortfolioGuard
 from ..risk.service import RiskService
 
@@ -100,12 +99,7 @@ async def run_cross_exchange_arbitrage(cfg: CrossArbConfig) -> None:
     engine = get_engine() if (cfg.persist_pg and _CAN_PG) else None
     if cfg.persist_pg and not _CAN_PG:
         log.warning("Persistencia habilitada pero Timescale no disponible.")
-    risk_mgr = RiskManager(risk_pct=0.0)
-    risk = RiskService(
-        risk_mgr,
-        PortfolioGuard(GuardConfig(venue="cross")),
-        risk_pct=0.0,
-    )
+    risk = RiskService(PortfolioGuard(GuardConfig(venue="cross")), risk_pct=0.0)
 
     async def maybe_trade() -> None:
         nonlocal position_sign, entry_edge
